refactored show preloader in movies and saved movies components
