+++ src/components/InfoPopup/InfoPopup.jsx
